uploadsys

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video"
+import User from "../models/User";
 
 export const home = async (req, res) => {
     const videos = await Video.find({});
@@ -6,8 +7,10 @@ export const home = async (req, res) => {
 }
 
 export const watch = async (req, res) => {
+    
     const { id } = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
+    
     if(video){
         return res.render("watch" , { pageTitle: "Watch", video})
     } else {
@@ -18,22 +21,37 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
     const { id } = req.params;
+    const {
+        user: { _id },
+    } = req.session;
     const video = await Video.findById(id);
     if (!video) {
         return res.render("404", {pageTitle: "Video not found!"})
-    } else {
-        return res.render("edit", { pageTitle: `Editing ${video.title}`, video });
-    };
+    } 
+    
+    if (String(video.owner) !== _id) {
+        return res.status(403).redirect("/")
+    }
+    
+    return res.render("edit", { pageTitle: `Editing ${video.title}`, video });
+
 }
 
 export const postEdit = async (req, res) => {
+    const {
+        user: { _id },
+    } = req.session;
     const { id } = req.params;
     const { title, description, hashtags } = req.body;
     const video = await Video.exists({_id: id});
     if (!video) {
         return res.render("404", {pageTitle: "Video not found!"})
-    } else {
-        await Video.findByIdAndUpdate(id, {
+    }
+    if (String(video.owner) !== _id) {
+        return res.status(403).redirect("/")
+    }
+
+    await Video.findByIdAndUpdate(id, {
             title, 
             description, 
             hashtags: hashtags
@@ -50,11 +68,15 @@ export const getUpload = (req, res) => {
 }
 
 export const postUpload = async (req, res) => {
+    const { user: {_id } } = req.session;
+    const file = req.file;
     const { title, description, hashtags } = req.body;
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title, 
             description,
+            fileUrl: file.path,
+            owner: _id,
             createdAt: Date.now(),
             hashtags: hashtags.split(",").map((word) =>(word.startsWith("#") ? word : `#${word}`)),
             meta: {
@@ -62,6 +84,9 @@ export const postUpload = async (req, res) => {
                 rating:0,
             },
         })
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/");
 
     } catch (error) {
@@ -71,6 +96,17 @@ export const postUpload = async (req, res) => {
 }
 
 export const deleteVideo = async (req, res) => {
+    const {
+        user: { _id },
+    } = req.session;
+    
+    const video = await Video.findById(id);
+    if (!video) {
+        return res.render("404", {pageTitle: "Video not found!"})
+    } 
+    if (String(video.owner) !== _id) {
+        return res.status(403).redirect("/")
+    }
     const {id} = req.params;
     await Video.findByIdAndDelete(id);
     return res.redirect("/")
